requirements.txt
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
google-generativeai==0.3.0
Pillow==10.1.0
instagrapi==2.0.0
apscheduler==3.10.4
